+++ bcbio/cwl/workflow.py
@@ -8,6 +8,7 @@ import toolz as tz
 from bcbio.pipeline import alignment
 
 ALWAYS_AVAILABLE = ["description", "resources"]
+STRING_DICT = ["config__algorithm__ensemble"]
 FLAT_DICT = ["config__algorithm__variantcaller"]
 
 def generate(variables, steps, final_outputs):